declare cleanupprocessingstage
