+++ src/homemade/game/model/GameModelLinker.kt
@@ -3,7 +3,6 @@ package homemade.game.model
 import homemade.game.fieldstructure.FieldStructure
 import homemade.game.loop.*
 import homemade.game.model.combo.ComboDetector
-import homemade.game.model.combo.ComboEffectVendor
 import homemade.game.model.spawn.SpawnManager
 import homemade.game.pipeline.GameUpdatePipeline
 import homemade.game.scenarios.GameOverScenario
@@ -50,34 +49,7 @@ class GameModelLinker(val structure: FieldStructure, val settings: GameSettings,
     }
 
     private fun updateStates() {
-        if (updater.hasCellChanges()) {
-            val comboPackTier = updater.comboPackTier()
-            ComboEffectVendor().addComboEffectsForTier(storedEffects, comboPackTier)
-            updater.takeChanges(spawner.markBlocksWithEffects(storedEffects))
-            trueState.selectionState.updateSelectionState()
-            updater.flush(trueState.configState.globalMultiplier)
-        }
-        if (state.getNumberOfBlocks() == structure.fieldSize) {
-            val multiplier = trueState.configState.globalMultiplier
-            if (multiplier > 1) {
-                modifyGlobalMultiplier(-multiplier)
-                updater.takeChanges(spawner.removeRandomBlocks())
-                updateStates()
-                println("multiplier consumed")
-            } else {
-                gameLoop.model.post(GameOver(1))
-                println("can't trade multiplier for blocks")
-            }
-        } else if (state.getNumberOfMovableBlocks() == 0) {
-            modifyGlobalMultiplier(BONUS_MULTIPLIER_FOR_BOARD_CLEAR + INITIAL_SPAWNS)
-            for (i in 0 until INITIAL_SPAWNS) {
-                ///requestSpawn()
-            }
-        }
+        //cleanup processing stage
     }
 
-    companion object {
-        private const val BONUS_MULTIPLIER_FOR_BOARD_CLEAR = 50
-        private const val INITIAL_SPAWNS = 2
-    }
 }
